improve generate registry entry cmd

+++ x/tokenregistry/client/cli/query.go
@@ -79,7 +79,7 @@ func GetCmdGenerateEntry() *cobra.Command {
 				return err
 			}
 			flags := cmd.Flags()
-			decimals, err := flags.GetInt(flagDecimals)
+			decimals, err := flags.GetInt64(flagDecimals)
 			if err != nil {
 				return err
 			}
@@ -177,7 +177,7 @@ func GetCmdGenerateEntry() *cobra.Command {
 				denom = baseDenom
 			}
 			entry := types.RegistryEntry{
-				Decimals:                 int64(decimals),
+				Decimals:                 decimals,
 				Denom:                    denom,
 				BaseDenom:                baseDenom,
 				IbcChannelId:             ibcChannelID,
@@ -206,7 +206,7 @@ func GetCmdGenerateEntry() *cobra.Command {
 		"The counterparty channel if this is an IBC token - used by clients when initiating send from a counterparty chain")
 	cmd.Flags().String(flagIbcCounterpartyChainID, "",
 		"The chain id of ibc counter party chain")
-	cmd.Flags().Int(flagDecimals, -1,
+	cmd.Flags().Int64(flagDecimals, -1,
 		"The number of decimal points")
 	cmd.Flags().String(flagUnitDenom, "",
 		"The denom in registry that holds the funds for this denom, ie the most precise denom for a token")
